+++ projects/mk-kit/src/lib/components/directives/index.ts
@@ -6,3 +6,4 @@ export * from './scrollspy';
 export * from './intersect';
 export * from './infinite-scroll';
 export * from './ripple';
+export * from './mask';